apps/pnl_checker/pyproject.toml
[project]
name = "pnl-checker"
version = "0.1.0"
description = "Live PnL validation tool — compares our calculations against Bybit"
requires-python = ">=3.11"
dependencies = [
    "gridcore",
    "bybit-adapter",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[tool.uv.sources]
gridcore = { workspace = true }
bybit-adapter = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pnl_checker"]
